+++ CHANGES.rst
@@ -94,4 +94,5 @@ VERSION 0.9.4
  - Now measures read/write times separately in diagnostic data
  - Performance improvement (no explicit loop over time-steps in a time-slice
    file needed)
+ - Enabling user-defined compression level for NetCDF4 output